requirements.txt
django==3.0
python-dotenv
flake8==3.7.0
six==1.16.0
pytest-django==4.5.2
coverage==6.4.1
sentry-sdk==1.6.0